Dedicate Uri parse tools to a separate library
