+++ core/common/CMakeLists.txt
@@ -1,3 +1,8 @@
+#
+# Copyright Soramitsu Co., Ltd. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+
 add_library(hexutil
     hexutil.hpp
     hexutil.cpp
@@ -34,3 +39,9 @@ target_link_libraries(mp_utils
     Boost::boost
     )
 kagome_install(mp_utils)
+
+add_library(kagome_uri
+    uri.hpp
+    uri.cpp
+    )
+kagome_install(kagome_uri)
